refactor moved pgp encrptyion to encryption package
